+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Made emulator patch version not impact machine root hash
 - Enabled line buffering for machine stdout in non-interactive mode
+- Changed stored machine configs from protobuf to JSON
 
 ## [0.15.1] - 2023-08-17
 ### Fixed